+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -24,6 +25,10 @@ Fixed for any bug fixes.
 Security in case of vulnerabilities.
 
 -->
+
+(changelog/unreleased)=
+
+## [Unreleased]
 
 ## [1.5.0] - 2022-11-25