+++ CHANGELOG.md
@@ -30,6 +30,9 @@ Security in case of vulnerabilities.
 
 ## [Unreleased]
 
+Supports Python 3.9 and above, from now on to make use of concise
+type-annotations.
+
 ## [1.5.0] - 2022-11-25
 
 ### Added